Refactor monolithic index.ts into route modules and shared libs (fixes #89)
Split the 4,480-line index.ts into 18 focused route modules and 18 shared
lib files. Each route module exports a FastifyPluginAsync and is registered
via app.register(). Shared utilities (prisma singleton, caches, TMDB/OMDB/RPDB
clients, auth, CORS, settings, metadata sync, watch events, Trakt polling, AI
recommendations) live in lib/ to eliminate duplication and enable independent
development of each feature area.
