+++ apps/api/src/lib/auth.ts
@@ -1,0 +1,26 @@
+import { createHash, timingSafeEqual } from "node:crypto";
+import type { FastifyReply, FastifyRequest } from "fastify";
+
+const API_TOKEN = process.env.API_TOKEN;
+
+const toSha256Digest = (value: string) => createHash("sha256").update(value).digest();
+
+export const verifyToken = async (request: FastifyRequest, reply: FastifyReply) => {
+  if (!API_TOKEN) {
+    request.log.error("API_TOKEN is not configured");
+    return reply.code(500).send({ error: "API token is not configured" });
+  }
+
+  const authHeader = request.headers.authorization;
+  if (!authHeader?.startsWith("Bearer ")) {
+    return reply.code(401).send({ error: "Unauthorized" });
+  }
+
+  const token = authHeader.slice("Bearer ".length).trim();
+  const tokenDigest = toSha256Digest(token);
+  const expectedTokenDigest = toSha256Digest(API_TOKEN);
+
+  if (!timingSafeEqual(tokenDigest, expectedTokenDigest)) {
+    return reply.code(401).send({ error: "Unauthorized" });
+  }
+};
