+++ apps/api/src/lib/tmdb-client.ts
@@ -1,0 +1,7 @@
+import { TmdbClient } from "../tmdb.js";
+import { getLanguageSetting } from "./settings.js";
+
+export const getTmdb = async (): Promise<TmdbClient> => {
+  const lang = await getLanguageSetting();
+  return TmdbClient.fromEnv(lang);
+};
