+++ apps/api/src/routes/health.ts
@@ -1,0 +1,7 @@
+import type { FastifyPluginAsync } from "fastify";
+
+const healthRoutes: FastifyPluginAsync = async (app) => {
+  app.get("/health", async () => ({ status: "ok", service: "api" }));
+};
+
+export default healthRoutes;
